+++ packages/grid/src/GridModel.ts
@@ -9,7 +9,6 @@ import {
 import memoizeClear from './memoizeClear';
 import GridUtils, { type Token } from './GridUtils';
 import { type CellRenderType } from './CellRenderer';
-import { type GridRangeIndex } from './GridRange';
 
 const LINK_TRUNCATION_LENGTH = 5000;
 
@@ -71,10 +70,6 @@ abstract class GridModel<
    * @returns Text for the specified cell
    */
   abstract textForCell(column: ModelIndex, row: ModelIndex): string;
-
-  tooltipForCell(column: GridRangeIndex, row: GridRangeIndex): string | null {
-    return null;
-  }
 
   /**
    * Get the character to replace text when truncated for a specific cell.
