+++ packages/grid/src/mouse-handlers/index.ts
@@ -8,5 +8,5 @@ export { default as GridRowTreeMouseHandler } from './GridRowTreeMouseHandler';
 export { default as GridScrollBarCornerMouseHandler } from './GridScrollBarCornerMouseHandler';
 export { default as GridVerticalScrollBarMouseHandler } from './GridVerticalScrollBarMouseHandler';
 export { default as EditMouseHandler } from './EditMouseHandler';
-export { default as GridTooltipMouseHandler } from './GridTooltipMouseHandler';
+export { default as GridTokenMouseHandler } from './GridTokenMouseHandler';
 export * from './GridSeparatorMouseHandler';
