+++ packages/iris-grid/src/mousehandlers/index.ts
@@ -8,5 +8,5 @@ export { default as IrisGridFilterMouseHandler } from './IrisGridFilterMouseHand
 export { default as IrisGridRowTreeMouseHandler } from './IrisGridRowTreeMouseHandler';
 export { default as IrisGridSortMouseHandler } from './IrisGridSortMouseHandler';
 export { default as PendingMouseHandler } from './PendingMouseHandler';
-export { default as IrisGridTooltipMouseHandler } from './IrisGridTooltipMouseHandler';
+export { default as IrisGridTokenMouseHandler } from './IrisGridTokenMouseHandler';
 export { default as IrisGridPartitionedTableMouseHandler } from './IrisGridPartitionedTableMouseHandler';
